Refactor royalties out of offerfacet
